Refactor: divide slack and email logic
