+++ src/modules/stores/aws.js
@@ -1,5 +1,5 @@
-const { aws: awsEnv, slackWebhookUrl: slackUrl } = require("../../../loadenv");
-const axios = require("axios");
+const { aws: awsEnv } = require("../../../loadenv");
+
 const logger = require("../logger");
 // Load the AWS-SDK and s3
 const AWS = require("aws-sdk");
@@ -84,7 +84,7 @@ module.exports.getS3Url = (filePath) => {
   return `${awsEnv.s3Url}${filePath}`;
 };
 
-module.exports.sendReportEmail = (reportUser, reportedEmail, report, html) => {
+module.exports.sendReportEmail = (reportedEmail, report, html) => {
   const reportTypeMap = {
     "no-settlement": "정산을 하지 않음",
     "no-show": "택시에 동승하지 않음",
@@ -115,27 +115,6 @@ module.exports.sendReportEmail = (reportUser, reportedEmail, report, html) => {
     if (err) {
       logger.info("Fail to send email", err);
     } else {
-      const data = {
-        text: `${reportUser}님으로부터 신고가 접수되었습니다.
-
-        신고자 ID: ${report.creatorId}
-        신고 ID: ${report.reportedId}
-        방 ID: ${report.roomId ?? ""}
-        사유: ${reportTypeMap[report.type]}
-        기타: ${report.etcDetail}
-        `,
-      };
-      const config = { "Content-Type": "application/json" };
-
-      axios
-        .post(slackUrl.report, data, config)
-        .then((res) => {
-          logger.info("Slack webhook sent successfully");
-        })
-        .catch((err) => {
-          logger.info("Fail to send slack webhook", err);
-        });
-
       logger.info("Email sent successfully");
     }
   });
